Server/requirements.txt
# packages in environment at /home/cho/anaconda3/envs/speaker:
#
# Name                    Version                   Build  Channel
cudatoolkit               10.1.243             h6bb024c_0  
django                    2.2.5                    py36_1  
djangorestframework       3.12.4                   pypi_0    pypi
keras                     2.2.1                    pypi_0    pypi
librosa                   0.8.1                    pypi_0    pypi
numpy                     1.16.6           py36h2d18471_3  
pyaudio                   0.2.11           py36h7b6447c_2  
python                    3.6.10               hcf32534_1  
pytorch                   1.3.0           py3.6_cuda10.1.243_cudnn7.6.3_0    pytorch
tensorflow                1.8.0                    pypi_0    pypi
